+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## RhaegalLib
 
-### Version 1.0 - 26/10/2019
+### Version 1.2.1 - 27/01/2020
 
-initial release 
+- Change the EVTX parser to more efficient parser, Thanks to https://github.com/omerbenamram/pyevtx-rs.
+- Added variables - this enable you to specify a variable instead a constant value (Environment variables, Other fields, etc). 
+- Added modifiers - allows you to perform processing on any field such as field length and regex match. 
+- Re-write "Event" class.
+- Added "returns" field - this allow you to return some fields instead of the full event log (check the WiKi for more details).
 
 ### Version 1.0.1 - 27/10/2019
 
@@ -12,12 +16,20 @@ initial release
 * Better Event log parser
 * Better Rhaegal rule validation
 
+### Version 1.0 - 26/10/2019
+
+initial release 
+
 ## Rhaegal
+
+### Version 1.0.2
+
+* Minor changes
+
+### Version 1.0.1 - 27/10/2019
+
+- Added logging
 
 ### Version 1.0 - 26/10/2019
 
 initial release
-
-### Version 1.0.1 - 27/10/2019
-
-* Added logging